Refactor order dashboard container
